Clarify that Perl 5.x with threading is needed
I installed Fast-Plast using the INSTALL.pl script. (Thanks so much for including that, by the way!) When we ran Fast-Plast, it failed complaining that it needed a threaded perl. So, I updated the README to indicate this dependency. 

I also changed "Perl 5+" to "Perl 5.x" so that people wouldn't think that they could install this using Perl 6 (confusingly a sister language to Perl 5 instead of the next version).